split bookshelf file into multiple file
